Add files via upload
Struggled with part b, the question on functional arrays - I think this was partially down to a subpar implementation of the Binary Trees, I think with more practice I could have done this in a much more efficient way than I did. I do want to redo this at some point, but being pushed for time, I have not had time to redo it this week, although the rest of the questions have helped with my understanding of how to use classes efficiently.